requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
flask>=2.0.0
jupyter>=1.0.0
notebook>=6.4.0
Pillow>=8.0.0
